+++ i18n/uk.go
@@ -2,13 +2,15 @@ package i18n
 
 var UK = Messages{
 	"syntax_error":             "Синтаксична помилка",
-	"unknown_person":           "Невідоме імʼя",
+	"unknown_person":           "Невідоме імʼя - %s",
 	"unknown_person_in_family": "В сімʼї %s немає людини з іменем %s",
-	"unknown_family":           "Невідоме прізвище",
+	"unknown_family":           "Невідоме прізвище - %s",
 	"create_family_after":      "Створити сімʼю %s після поточної %s",
 	"create_family_at_end":     "Створити сімʼю %s в кінці цього файлу",
 	"create_family_file":       "Створити сімʼю %s у новому файлі",
 	"change_name_from_source":  "Замінити на %s - дитину %s",
 	"duplicate_count_of_name":  "Неоднозначне імʼя. У цій сімʼї %d імені %s. Використайте унікальний варіант імені одного з них",
 	"child_of_source":          "дитина подружжя %s",
+	"child_without_relations":  "%s %s не має відносин",
+	"create_child_relation":    "Створити сімейні відносини",
 }